Bug 565143 - [Passage] floating licensing
Move Port to LBC base
