--- tests/org.eclipse.passage.lbc.server.tests/.classpath
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
- Copyright (c) 2020 ArSysOp and others
- This program and the accompanying materials are made available under the
- terms of the Eclipse Public License 2.0 which is available at
- https://www.eclipse.org/legal/epl-2.0/.
-
- SPDX-License-Identifier: EPL-2.0
-
- Contributors:
-     ArSysOp - initial API and implementation
--->
-<classpath>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
-	<classpathentry kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
-	<classpathentry kind="src" path="src">
-		<attributes>
-			<attribute name="test" value="true"/>
-		</attributes>
-	</classpathentry>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
