--- tests/org.eclipse.passage.lbc.server.tests/.settings/org.eclipse.pde.core.prefs
@@ -1,3 +1,0 @@
-eclipse.preferences.version=1
-pluginProject.extensions=false
-resolve.requirebundle=false
