--- tests/org.eclipse.passage.lbc.server.tests/META-INF/MANIFEST.MF
@@ -1,11 +1,0 @@
-Manifest-Version: 1.0
-Automatic-Module-Name: org.eclipse.passage.lbc.server.tests
-Bundle-ManifestVersion: 2
-Bundle-SymbolicName: org.eclipse.passage.lbc.server.tests
-Bundle-Version: 1.0.0.qualifier
-Bundle-Name: %Bundle-Name
-Bundle-Vendor: %Bundle-Vendor
-Bundle-Copyright: %Bundle-Copyright
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Require-Bundle: org.junit;bundle-version="4.13.0"
-Fragment-Host: org.eclipse.passage.lbc.server
