--- tests/org.eclipse.passage.lbc.server.tests/OSGI-INF/l10n/bundle.properties
@@ -1,22 +1,0 @@
-#Properties file for org.eclipse.passage.lbc.server.tests
-###############################################################################
-# Copyright (c) 2020 ArSysOp and others
-#
-# This program and the accompanying materials are made available under the
-# terms of the Eclipse Public License 2.0 which is available at
-# https://www.eclipse.org/legal/epl-2.0/.
-#
-# SPDX-License-Identifier: EPL-2.0
-#
-# Contributors:
-#     ArSysOp - initial API and implementation
-###############################################################################
-Bundle-Name = Passage LBC Server Tests
-Bundle-Vendor = Eclipse Passage
-Bundle-Copyright = Copyright (c) 2020 ArSysOp and others.\n\
-\n\
-This program and the accompanying materials are made\n\
-available under the terms of the Eclipse Public License 2.0\n\
-which is available at https://www.eclipse.org/legal/epl-2.0/\n\
-\n\
-SPDX-License-Identifier: EPL-2.0\n\
