--- tests/org.eclipse.passage.lbc.server.tests/build.properties
@@ -1,18 +1,0 @@
-###############################################################################
-# Copyright (c) 2020 ArSysOp and others
-#
-# This program and the accompanying materials are made available under the
-# terms of the Eclipse Public License 2.0 which is available at
-# https://www.eclipse.org/legal/epl-2.0/.
-#
-# SPDX-License-Identifier: EPL-2.0
-#
-# Contributors:
-#     ArSysOp - initial API and implementation
-###############################################################################
-source.. = src/
-output.. = bin/
-bin.includes = META-INF/,\
-               .,\
-               OSGI-INF/,\
-               about.html
